Add files via upload
in this case, it's the first update. the game must be based on the state vectors of the Bloch sphere representing each station, to calculate the matrix of the rotationals logic gate to be able to go from one station to another. the day can also be based on the amplitude (identifiable through the volume of each basic state) and phase (identifiable through the color) of each station.